Fix in LookupComponent. New api call nameservice.getEntryCount() that asks all superpeers and aggregates the number of nameservice entries in the system.

+++ src/de/hhu/bsinfo/dxram/lookup/overlay/OverlayPeer.java
@@ -281,8 +281,8 @@ public class OverlayPeer implements MessageReceiver {
 	 * Get the number of entries in name service
 	 * @return the number of name service entries
 	 */
-	public long getNameserviceEntryCount() {
-		long ret = 0;
+	public int getNameserviceEntryCount() {
+		int ret = 0;
 		Short[] superpeers;
 		GetNameserviceEntryCountRequest request;
 		GetNameserviceEntryCountResponse response;
@@ -302,8 +302,6 @@ public class OverlayPeer implements MessageReceiver {
 				ret += response.getCount();
 			}
 		}
-
-		// TODO: Aggregate counts and add command to service
 
 		return ret;
 	}
